+++ app/src/main/java/com/bn/todo/data/model/TodoFilter.kt
@@ -1,7 +1,7 @@
 package com.bn.todo.data.model
 
 data class TodoFilter(
-    val listId: Int,
+    val listId: Long,
     val showCompleted: Boolean
 ) {
     var name: String? = null
